requirements.txt
streamlit>=1.30.0
langchain>=1.2.10
langchain-ollama>=0.1.0
python-dotenv>=1.0.0